Add finders
Finders are objects that understand a specific type of packet. It is initialised
with two strings which define a pattern matching system. The first string contains
exact characters found in the packet. The second contains match types:

e: exact match
u: uppercase letter
l: lowercase letter
n: number 0-9
h: hexadecimal number

The finder looks at the decoded packet to see if it matches. If it does, it supplies
a more constrained set of possible bytes to the deconvolver, which can then be run
again to give a better solution to the packet.