subpage squasher: Takes all the pages found at a specific page number
eg 100 and builds a list of the unique subpages by computing hamming
distance between each pair of pages, and also looking at the subpage
numbers. When it has multiple copies of each subpage it takes the mode
of each byte (ie the most common value) to "squash" them into a single
page. Then it can generate a binary string or html.