Add fragment testers.
These area bit like the packet finders. They test for know
data. But they work on sub-regions of a single subpage.

For example, a fragment could test for the BBC logo.

The fragment contructor takes an array containing the bytes
to look for, and a (x,y,w,h) describing the rect to search
on pages.

fragment.py also contains FragmentBuilder which searches
in a large number of subpages to find common fragments,
and then dumps them. Then you can copy paste that output
back into the source code to define a new fragment.