Tweak finders and output individual files
This is a big change that does two things. Finders are now all
tested and the best match is used. This means you don't need to
specify which ones to use any more.

That change was a good time to bring in the individual output code
becuase it breaks the old code anyway. The new method writes a t42
file for each vbi file. You must make a directory structure for
the data like this:

0001/vbi <- vbi files go in here.
0001/t42 <- t42 files will be put in here.

Then you run:

./vbi.py 0001/

vbimp.py is no longer needed so it is deleted.