A big speedup from this change.
Two main differences:

The stuff with the guess convolver is moved to a separate class.
It is capable of re-convolving only the bytes that have changed
instead of recalculating the whole thing. This gives approx 2x
speed-up.

The possible-byte masks are now precalculated once and then
fetched from a look-up table. This gives another 2x speed-up.
Also all the byte lists are now sets which can be combined more
quickly.