One more fix for guess.py
When generating the guess masks the input signal is partitioned
into bytes. That partitioning is now done with a faster method.
But it had a bug. The upper limit must have the bitwidth added
before casting to int to prevent a rounding error.

With this change the whole process should now produce the same
results as before, if it is given te same input parameters.