Don't use Pool.starmap
This doesn't lazy evaluate the iter. It tries to load the whole
thing before processing any of it. This is bad when you have 16G
of VBI. Instead we have to use Pool.imap and unpack the tuple in
the callable.

Also because this is Python 3 we can just use map instead of
itertools.imap because it *does* lazy evaluate the iter now.