Refactor packet types
This changes packet types to internally store the encoded teletext
data. It is decoded on demand instead of at object creation.

Elements no longer use descriptors. They are created on demand and
store a view into the data. Their properties are also created on
demand. This makes it a bit easier to write them. You only need
to define get() to return a dict of all properties and set() to
take optional kwargs for all properties. The rest is done by the
wrappers.