More refactoring in packet.py and coding.py
Coding is now done exclusively with array indexing. It also supports
array arguments.

The Packet subclasses are gone. The row field on a packet can be
changed which leads to the object having the wrong type for its data.
Instead there is now just one Packet type which exposes everything.