Implement statistics
For packet handlers histograms of magazines and rows are available.
For deconvolve, percentage of rejected lines is also available.

Statistics are shown on the progress bar. This commit turns off
progress bars by default for all commands except for deconvolve.
However, it adds the ability to turn them on for most commands.

By default no statistics are shown even if you enable a progress
bar, except for deconvolve which shows rejects and magazine
histogram by default.

Fixes #14.