Add min-duplicates and pages options to squash
Minimum duplicates is the minimum number of duplicates required
for a subpage to be considered for squashing. Less than 3 will
produce bad quality results as there can never be a 'most common'
result from two or less choices. However it is necessary to find
all subpages, especially on short recordings.