Re-factor Line and add OTA slicing
This adds a new sub-command "slice". It works like "deconvolve"
except that it works on VBI recordings taken live. The regular
deconvolver can't handle these. This simple implementation uses
slice/threshold which is not optimal, but works well enough to
see recognisable pages.

This also simplifies the teletext detection and removes one of
the rules to allow OTA VBI to be correctly detected.