Fix pipe errors on Windows
Windows allows oyu to seek on pipes and reports the current amount
of data in the buffer. So the file chunker would stop after reading
that much data, believing it to be a regular file. This is fixed by
checking if the file is a pipe, and raising ioUnsupportedOperation,
which is what happens if you try to seek on a pipe in Linux.