Make ctrl-c behave nicely with multiprocessing
This adds a context manager, SigIntDefer. Code running inside this
context cannot be interrupted by ctrl-c, but it can check if ctrl-c
has been pressed and then exit at an appropriate time. The context
manager restores the original signal handler when exiting and calls
it if a signal was caught during execution. It can also restore the
original signal handler after some number of signals are received
(default 1), meaning that pressing ctrl-c twice will still behave
like a normal ctrl-c.

This is used in itermap multiprocessing to ensure that we always
deque exactly as many work items as we queue. If this is not done
then the program deadlocks at exit and can leave zombie processes
running.