Implement pagination as a decorator
Also implement detecting invalid argument combinations, such as
filtering out row 0 when paginating.

The paginated decorator has two options:

"always" means that the "-P" option will not be available on the
command line and the argument will not be passed to the ultimate
handler. This for commands like "html" which need full pages.
Default is False.

"filtered" means that the extra options for pages and subpages
will be available on the command line and passed to the handler.
Some handlers need to do their own thing with these arguments.
Defaults to True.

Fixes #29