Implement a better method to choose when subpages gets two of the same row
Previously this would just take the last one it saw. It now chooses based on
the number of 0x80 bytes in each packet. This byte is parity 0 and a large
number of them incidate that the packet is empty/filler. This works well for
selecting the most "interesting" display packet when two are available.

The downside of this is it isn't what a regular decoder would do. That would
have the original behaviour of showing the last received packet.