Add Teletext to Unicode 13.0 mappings
These are taken from:

https://www.unicode.org/L2/L2019/19025-terminals-prop.pdf

The G0 mapping corrects many of the special characters not previously
handled and changes a couple of them to similar looking but different
code points. Overall this is an improvement.

The G1 mapping is very different to the private use scheme previously
used and will require further refactoring in order to use it.