More refactoring of MP code
There is now a lot of unit tests which should catch most possible
ways you can deadlock the MP code. It is also no longer necessary
to catch ctrl-c/SIGINT, as the main process does not need to keep
track of items in/out the queue in order to shut down cleanly.