Try to handle the case where multiprocessing kills all the workers
When you yield from inside a with: context, exceptions can happen
outside the context. This can cause the python interpreter to exit
before exiting the with: context. When the context object gets
deleted, the exit handler runs. But this happens after any functions
registered with atexit. An example is shown below:

    import atexit
    import random

    def exitfunc():
        print("atexit runs")
    atexit.register(exitfunc)

    class Context(object):
        def __enter__(self):
            return self
        def __exit__(self, *args):
            print("exiting context")

    def gen():
        with Context() as c:
            while True:
                yield "hello"

    g = gen()
    next(g)

In this example the atexit handler will run before the context
exits.

This is a problem because multiprocessing registers an atexit
handler that kills all daemon subprocesses with SIGTERM. This
means that sometimes out pool context will normally exit with
the workers still alive, but sometimes it will exit when they
have already been terminated.

The exact circumstance where this happens is subject to a race
condition: it depends exactly where in the program the SIGINT
(or other exception) happens.

This patch tries to make the exit handler work correctly no
matter what state the program is in when it is called.