Switch mp to use a pipe per subprocess instead of shared queue
This has far fewer corner cases that can cause deadlocks. If a worker
crashes, the pipe will close and raise an error on reading. This removes
the need for complex events and timeouts, and the pipes do not have to
be drained before workers can stop. That in turn means we don't have to
worry if multiprocessing terminated the workers before we could. We only
need to join workers if we're doing normal cleanup but not exiting the
program. So we don't really need to care about ctrl-c at all any more,
other than telling workers to ignore it.

All of this means the code is far simpler. And as a bonus, it is about
10% faster too!