Implement multithreaded worker processes
Each worker creates two additional threads. One to read work and
the other to send back results. The processes and threads are not
daemon so they will never be automatically closed. They are told
to close by sending a work item with number < 0. As soon as this
is received, the reader thread dumps all waiting work and puts
the exit message on the work queue. When the main thread receives
this work item it empties the result queue and puts the exit
message on it. Then finally the writer threads receives the exit
message and writes it back in to the pipe. Then the main thread
joins the other two.

Back in the main process we raise a ChildProcessError when we
get back the exit message. This should never happen in normal
operation, but may if a worker crashed in certain ways.

A very big advantage of this method is that all threads can block
on their respective operations. No thread needs to time out to
check for a quit event. This means that no CPU is used when a
pool is idle (important for GUI.)

Also, taking full control of tearing down the processes means
that there is no way for system/GPU resources to be leaked.
We always exit cleanly from workers whenever it is possible.