Ensure shutdown with atexit
Multiprocessing tries to join any remaining processes with an atexit
handler. Due to with/yield issues, this can run before the context
handler for the pool. Ensure that we shutdown before this happens by
registering our own atexit handler. Since we register it after
multiprocessing, it will run before. On normal shutdown it is
unregistered.