Don't allow user to record to the VBI device
If the user tries to use a path beginning with "/dev/vbi" as the
output file, refuse to proceed and print an error message.

This catches the mistake where you do something like:

    teletext record /dev/vbi1

This will read from vbi0 and write to vbi1. This is bad as writing
to the capture device seems to make some cards hang up.