Make commands that deal with t42 handle padding packets
Padding packets are empty packets inserted to make the packet offset
in the output match the offset in the input. They contain all zeros.
This patch makes t42 related commands handle them as appropriate.

Commands that re-order the output are made to filter out padding.
Commands that keep the packets in order do not filter padding. The
filter command filters padding by default, but can be told to keep
it.