Introduce a new scan command
This simplifies the task of scanning a full tape recording for
individual services. It squashes together headers and BSDP packets
from some number of frames and then prints a summary of the result.
It also shows an indication of which lines contain active teletext.

Specify number of lines per frame with -l and number of frames to
average with -f.