Add pagination to deconvolve command and line count to VBI viewer
This allows you to select a specific row on a specific page when
deconvolving, and pipe the result into the vbiview for adjustment.
Useful when you have a line of text that is just so mangled you
can't figure out what it is supposed to say.

Use it like this:

    teletext deconvolve samples.vbi -p 100 -r 5 -o vbi - | teletext vbiview -p -n 1

Now wait for the deconvolve to find the packet, and it will be displayed
in vbiview. Use the mousewheel to adjust offset to try to get a better
deconvolution. Press n to go to the next instance of this packet.

You can of course you any combination of filters you want on the
deconvolve command.