VBI/opencl: 2 phase opencl min
Split the min process into two phases; and 'minpar' sets of patterns.
The first phase takes 'minpar' results for each char and boils it
down into a minimum for that set.  The second phase then reduces
it further to the result of minimum per character.

minpar is currently set at 256 which seems like a nice round number
and is getting pretty good results (926 lps) for me.

Signed-off-by: Dr. David Alan Gilbert <dave@treblig.org>