VBI/opencl: minpar heuristic
Use 1024 for the larger pattern sets and 256 for the smaller ones.
On my hardware, 1024 certainly gets the min down from about 3.8ms to
about 1ms.  But that's a bit much for the 8k pattern sets and it's
marginal for the 32k set; so only use 1k for the big sets.