VBI/opencl: Add a missing wait
Add a missing wait() at the end of the queued chain.
Not sure if this has actually caused any corrupt data but I guess
it's possible.
Foudn when trying to turn profiling on, only broke with the Mesa
set, not the ROCm set for some reason - but I think Mesa is right.