Clustering: Cluster by row number and add n_lines option to histogram
Clustering by row number helps a lot because the contents of a row
doesn't change very often. It doesn't increase the number of output
clusters all that much, but reduces the number of mixed clusters.

Histogram gets an option to set the number of lines in the output.
This is useful for rendering a cluster histogram with n_lines=1.