Subpage: Make to_tti return bytes
The TTI format isn't really a plain text format. It nearly is, but
it has requirements for certain types of newlines and escaping that
aren't really compatible with writing strings to files. Therefore
generate the data as bytes, and write it to a file opened with 'wb'
in the example.